requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pymongo>=4.6.0
python-dotenv>=1.0.0
requests>=2.31.0
spotipy>=2.23.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
python-jose[cryptography]
passlib[bcrypt]
multipart